requirements.txt
wheel
Pillow==8.0.0
numpy==1.18.1
umsgpack
Cython
tqdm
scikit-image
scikit-learn 
inplace-abn==1.1.0
pyyaml
flake8
setuptools==40.8.0
tensorboardX==2.0
torch==1.4.0
torchvision==0.5.0
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17


